+++ NEWS.md
@@ -1,6 +1,8 @@
 # palaeoverse (development version)
 
 * Added age arguments (max_ma/min_ma) to bin_time (#106)
+* Added warning to bin_time (#104)
+* Improved link accessibility (#88)
 
 # palaeoverse 1.2.1